requirements.txt
Selenium
pyautogui
pandas
pymongo
python-dotenv
pandas
fpdf
flask
flask_pymongo
certifi